Assign lable name for legends using DataFrame or Series name
Check for a name attribute on the input and automatically set that as
the label if it exists for Pandas DataFrames and Series. If the user
specifies a label in the call to kdeplot it will take precedence over
the name attribute. If no label is provided and there is no name
attribute then a default value of "kde" is assogned.