More work on kdeplot and legends
Did more work on legend handling in kdeplot.

Made changes to label and legend handling in distplot. added keyword
kde_label (defaults to False). If it's set to True, then
kde_kws["label"] is set to "kde", and the KDE in the resulting plot
will be labeled "kde".

Added more complete description of how to use kdeplot labels to
notebook. Added  example of vertical kdeplots.