Use matplotlibrc file to force Agg backend in tests
This will avoid errors in tests that need a rendered plot. It might not
be the best way to force the Agg backend, but seems like the obvious way
to do it. I'd rather not keep the file in the root directory of the package
because lots of times I plot from within that directory and it would lead
to confusing issues.