Set rcmod.py to create translucent legend frames
As of 1.4.1, matplotlib includes a key for the default transparency of
legend frames: "legend.framealpha". *Personally,* I prefer the look of
this when a plot doesn't have much free space for the legend to avoid
overlap with data, but I leave it to others to accept or not. I also
don't know how to best handle compatibility with earlier versions of
mpl.