Fix bug in Grid legends
Previously, `add_legend` was passing a list of handles for the plots
that had been drawn and the list of all hue levels to the matplotlib
legend code. Because matplotlib just zipped these and didn't raise
when they were of differenet lengths, this meant that sometimes the
right colors but wrong labels were used.

This isn't globablly optimal, because I think the behavior we want is
to show the hue levels that could have existed in the plot (because
they are in `hue_order` but don't, because they weren't in the DataFrame.

However, doing that is a big tricky because we'll have to make a proxy
artist, you can't just pass `None` or something similar as a handle.