Compute linkage correctly when fastcluster is not installed.
When fastcluster is not installed, _calculate_linkage_scipy is used to
compute linkages. It contained an erroneous call to squareform().
distance.pdist() already returns a condensed distance matrix that is
required as input for the linkage() function. Converting it to a matrix
with squareform() is possible and linkage() will accept the resulting
matrix, but consider it to be a matrix with vectors of observations and run
distance.pdist() again on it. Thus, we got a linkage based on distances
between distances.