Fix #719 by converting the format templates in FacetGrid.set_titles() to Unicode.
This commit adds the function ``to_utf8`` to utils.py which reliably
converts Python objects to unicode strings both in Python 2.7 and 3.x. This
function is used in ``FacetGrid.set_titles()`` to ensure that the string
format templates used in the plot titles are always unicode strings.

The function ``to_utf8`` is somewhat similar in spirit to the one suggested
in #724. That commit, however, worked only in Python 2.7, and broke Python 3.x
compatibility.