Fix legend_out with GUI backend
There was a bug that FacetGrid.add_legend() with legend_out=True did not
work when the backend was "GUI"-based ones such as Qt5Agg and
NbAgg (checked with matplotlib 2.0.2).  Note that it was working
correctly with Agg backend.  This is likely because pyplot.draw calls
canvas.draw_idle() which may not force re-draw [1].

[1] https://github.com/matplotlib/matplotlib/blob/v2.0.2/lib/matplotlib/pyplot.py#L691