ENH: adv subtitled legends
Add in `align_legend_subtitles` function to change existing legends into subtitled legends. This is the main entry point, and uses the following helper functions to achieve a flexible function to create subtitled/sectioned legends.

`gather_legend_layout` is a helper function to determine the current subtitled set up of a legend, as generated by seaborn's relational plots (legends with invisible or size=0 handles to denote subtitles). `create_legend_layout` is a helper function that creates a label -> entry dictionary (where an entry is a handle/label HPacker from inside of an existing legend) and then creates a subtitled layout for a new legend as per user specifications. `iter_entries` iterates over all of the HPacker instances that have a DrawArea and a TextArea inside of them and is used by the two aforementioned functions.