To violinplot(), a new "fixed" option was added for the scale parameter.
In the case of making multiple violinplots on separate but related data series,
it may be desirable to have the plots share the same maximum counts scaling parameter value.
By creating a new "fixed" option for the scale parameter, and populating it (default is 200),
violinplots can be made that share the same scaling parameter while preserving plot features
in the manner of "counts" mode.

1. declare self.fixedmaxcount = fixedmaxcount in _ViolinPlotter
2. in estimate_densities add : elif scale == "fixed":
3. add scale_fixedmaxcount
4. added some documentation