Added option for ci to be a callable.
Added the option for ci to be a callable in relational.py. If ci is a callable, it should take a list of values representing multiple observations of the ``y`` variable at the same ``x`` level and return a list [lower_bound, upper_bound] where lower_bound is the lower value for the shaded region and upper_bound is the upper value for the shaded region. ci can also return a different length-two array-like object. A ValueError exception is raised if ci is callable but doesn't satisfy these conditions. I also modified the documentation to reflect this change.