Working "hello world" for non-notebook user
I was pretty confused by the intro docs. I wanted a "hello world" ASAP, but `sns.pairplot` was doing nothing. It just returns an object.

It turns out that behind the scenes, seaborn updates the pyplot context, and to see the plot, I needed to use `.show()`.

The current docs seem to assume the user is in a notebook environment. I do not believe this applies to everyone.