Add Grid.figure to (eventually) replace Grid.fig (#2639)
* Add Grid.figure to (eventually) replace Grid.fig

The advantage of this change is that you will be able to do

    obj = sns.<func>()
    obj.figure.<method>

regardless of whether  is figure-level or axes-level.

The `Grid.fig` methods are being soft-deprecated: discouraged in the
attribute documentation but not (currently) issuing a warning when used.

* Update docs, tests, and release notes

(cherry picked from commit 091f4c0e4f3580a8060de5596fa64c1ff9454dc5)