TST: Make test_index_alignment_between_series less brittle
test_index_alignment_between_series fails on Windows but passes on Linux.
The reason is that the indices of the expected Series are created using
the default dtype of an integer np.arange which is int64 on Linux but
int32 on Windows. Now the expected index is created the same way as the
x/y-indices and hence results in the same dtype.